Change model of Repeatable and Completable
