+++ src/main/java/seedu/address/model/task/Interval.java
@@ -1,0 +1,13 @@
+package seedu.address.model.task;
+
+/**
+ * Recognised Intervals of Interval.
+ */
+public enum Interval {
+    NONE,
+    DAILY,
+    WEEKLY,
+    FORTNIGHTLY,
+    MONTHLY,
+    YEARLY,
+}
