--- src/main/java/seedu/address/model/task/Recurrable.java
@@ -1,16 +1,0 @@
-package seedu.address.model.task;
-
-public interface Recurrable {
-
-    /**
-     * Returns the recurrence interval.
-     * @return Recurrence interval.
-     */
-    Recurrence getRecurrence();
-
-    /**
-     * Sets the Recurrence interval to specified level.
-     * @param recurrence Level of Recurrence.
-     */
-    void setRecurrence(Recurrence recurrence);
-}
