--- src/main/java/seedu/address/model/task/Recurrence.java
@@ -1,13 +1,0 @@
-package seedu.address.model.task;
-
-/**
- * Recognised Intervals of Recurrence
- */
-public enum Recurrence {
-    None,
-    Daily,
-    Weekly,
-    Fortnightly,
-    Monthly,
-    Yearly,
-}
